requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
flask[async]==3.0.0
requests==2.31.0
pyngrok==7.0.0
openai==1.3.0 